+++ workspaces/orchestrator/plugins/orchestrator-form-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @red-hat-developer-hub/backstage-plugin-orchestrator-form-react
 
+## 2.6.2
+
+### Patch Changes
+
+- Updated dependencies [309547d]
+  - @red-hat-developer-hub/backstage-plugin-orchestrator-common@3.5.2
+  - @red-hat-developer-hub/backstage-plugin-orchestrator-form-api@2.5.2
+
 ## 2.6.1
 
 ### Patch Changes